+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/jupiter/annotation/Spending.java
@@ -14,7 +14,7 @@ public @interface Spending {
 
   String description();
 
-  double amount();
+  int amount();
 
   CurrencyValues currency() default CurrencyValues.RUB;
 }
