+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/model/SpendJson.java
@@ -17,7 +17,7 @@ public record SpendJson(
         @JsonProperty("currency")
         CurrencyValues currency,
         @JsonProperty("amount")
-        Double amount,
+        int amount,
         @JsonProperty("description")
         String description,
         @JsonProperty("username")
